requirements.txt
# UI
streamlit==1.41.1

# Google Sheets
gspread==6.1.4
google-auth==2.37.0
google-auth-oauthlib==1.2.1

# AI/LLM
anthropic==0.42.0

# ブラウザ自動化
playwright==1.49.1

# スクレイピング
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0

# RSS解析
feedparser==6.0.11

# ニュース収集
gnews==0.3.7

# Excel処理
openpyxl==3.1.5

# ユーティリティ
python-dotenv==1.0.1
pydantic==2.10.3
Pillow==11.0.0
